web/backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# GPO Analyzer dependencies (from original)
beautifulsoup4==4.12.2
pandas==2.1.3
openpyxl==3.1.2
lxml==4.9.3

# API utilities
pydantic==2.5.2
pydantic-settings==2.1.0

# Authentication (Entra ID / Azure AD)
python-jose[cryptography]==3.3.0
httpx==0.25.2
cachetools==5.3.2
